Moved HTTP proxy to generic proxy as SOCKS is also supported
